added new get paths to markets
